Search and View Results
- Converted SearchDetails component to Details page
- Changed all axios requests to use fetch api
- Changed nodemailer request link to heroku link
- On clicking VIEW, docId is used by Details page to get data
